Refactor ToggleCheckbox into folder for reusuable components
